VISTA-414 - UI correction (Deposit and Concession money/percent value behaviour)

+++ vista-shared/vista-domain/src/main/java/com/propertyvista/domain/financial/offering/Concession.java
@@ -13,6 +13,7 @@
  */
 package com.propertyvista.domain.financial.offering;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import javax.xml.bind.annotation.XmlType;
@@ -20,6 +21,7 @@ import javax.xml.bind.annotation.XmlType;
 import com.pyx4j.commons.LogicalDate;
 import com.pyx4j.entity.annotations.Detached;
 import com.pyx4j.entity.annotations.Editor;
+import com.pyx4j.entity.annotations.Editor.EditorType;
 import com.pyx4j.entity.annotations.Format;
 import com.pyx4j.entity.annotations.JoinColumn;
 import com.pyx4j.entity.annotations.Length;
@@ -146,7 +148,7 @@ public interface Concession extends IVersionedEntity<ConcessionV> {
         @ToString(index = 2)
         @Format("#0.00")
         @MemberColumn(name = "val")
-        IPrimitive<Double> value();
+        IPrimitive<BigDecimal> value();
 
         @MemberColumn(name = "cond")
         IPrimitive<Condition> condition();
